+++ pipelines/publish/helpers/create_version.py
@@ -3,26 +3,12 @@ import pathlib
 import requests
 
 from gcip2 import Predefined
+from gcip2.vault import SecretsHandler
 
 
 def create_version_tag(version: str):
-    vault_url = "https://vault.pivlab.space"
-
-    vault_token = requests.post(
-        url=f"{vault_url}/v1/auth/pivlab/login",
-        data={
-            "jwt": Predefined.CI_JOB_JWT.getenv(),
-            "role": "rnd",
-        },
-    ).json()["auth"]["client_token"]
-
-    token = requests.get(
-        url=f"{vault_url}/v1/rnd/data/infra/accounts/gitlab/ci",
-        headers={
-            "X-Vault-Token": vault_token,
-            "X-Vault-Namespace": "rnd/",
-        },
-    ).json()["data"]["data"]["token"]
+    client = SecretsHandler("gitlab-token")
+    token = client.fetch()["token"]
 
     gitlab_url = Predefined.CI_API_V4_URL.must()
     project_id = Predefined.CI_PROJECT_ID.must()